convert schema to a mutable constructor
